Generalize to get ready for encoding.
This involves quite a bit of refactoring, and adding of new functions
in cbor_decode, as well as extracting some functions into a cbor_common
library.

This adds backups as part of the state variable.

This also adds checking for whether the whole input has been decoded.

This also fixes a number of large and small issues in cddl_gen.py:
 - Better respect CDDL's operator precedence
 - Better handling of skipping levels in the type hierarchy.
 - General code cleanup.

Signed-off-by: Øyvind Rønningstad <oyvind.ronningstad@nordicsemi.no>