extensions.cmake: Convert cddl_gen call to execute_process
instead of add_custom_command, so it is run immediately when
called in Cmake.

It looks like this function will primarily be used once in a
while, not as a regular part of the build process. Making the
call happen immediately makes it less intrusive to add or
remove it from the build process, e.g. via a Cmake flag.

Signed-off-by: Øyvind Rønningstad <oyvind.ronningstad@nordicsemi.no>