tests: scripts: Add test code and modify vectors and cddl
Some of the vectors needed changes to decode correctly.
Some was changing the order because of maps, some was fixing errors.

Change the CDDL to be more parsable.
Changing the order of maps.
Add "missing" element
Add CDDL for COSE_Sign1

Add new tests to tests/test.sh and to the Github workflow

Signed-off-by: Øyvind Rønningstad <oyvind.ronningstad@nordicsemi.no>