Apply small union improvement.
Make all calls to functions in cbor_decode.c restore the previous
state if they fail. This means the union_elem_code() isn't always
needed.

Unions are decoded as follows:

 - Save state
 - Attempt first element in union
 - If failed, restore state and try next element
 - etc.

union_elem_code() is a C function that is called in the generated
code, between each element in a union to restore the state.

Signed-off-by: Øyvind Rønningstad <oyvind.ronningstad@nordicsemi.no>