cbor_common.c: Utilize the unused 0th backup.
current_backup starts at 0, and is incremented to 1 the first time a
backup is called for. This change uses the backup at current_backup - 1
instead of current_backup, so the 0th backup can be used.

Also, add support for state->backups being NULL. Which is equivalent to
having a backup list of size 0.

Signed-off-by: Øyvind Rønningstad <oyvind.ronningstad@nordicsemi.no>