cddl_gen.py: Add additional parsing of labels
To better support interaction with / and =>

The operators ':' '/' and '=>' are handled in a special case, in which
their regex contains all other regexes. To avoid recursion, "all other
regexes" does not include the regexes for the above operators.

This commit adds a "non-special" regex for ':' so that certain CDDL
descriptions where ':' is used with the other operators work, and
don't throw exceptions.

This also adds tests for usage of ':' and '/' together.

Fixes issue #64.

Signed-off-by: Øyvind Rønningstad <oyvind.ronningstad@nordicsemi.no>