cddl_gen.py: Various changes to the logic for naming in code generation
The logic for giving names to variables, members, and types in the
generated code.

The old logic cause some ambiguity when parsing manifest14.cddl.
This fixes that, and does a few things to make some names shorter.

This introduces the base_stem variable which contains the top-level type
from the CDDL. This is used together with the existing id_prefix when
generating ids.

This also tweaks the existing id_prefix system e.g. by resetting the id
when creating a new C type which doesn't have a corresponding type in
the CDDL.

Signed-off-by: Øyvind Rønningstad <oyvind.ronningstad@nordicsemi.no>