zcbor.py: Add support for 64 bit integers in code generation
Whenever an integer (UINT, NINT, or INT) has a specified value,
or max or min value, that requires 64 bits to represent, they
gain a uint64_t or int64_t type, and use the 64 bit decoding
or encoding functions.

The default is still 32 bits, but this can now be changed via
the --default-bit-size cli option.
when --default-bit-size is 64, all integers will be 64 bits,
regardless.

Transition a number of uint32_t variables to either size_t or
uint_fast32_t. Some unused variables were removed in zcbor.py

Run all tests with --default-bit-size 64 when running on
native_posix_64.

Signed-off-by: Øyvind Rønningstad <oyvind.ronningstad@nordicsemi.no>