zcbor.py: Fix bug where the union_int optimization decodes twice
When a union contains another type, which is disambiguated, but which
will be decoded with its own function, the first integer would be
attempted decoded twice, once to populate the choice enum, and once
in the type's own function. This change reports such cases as
not disambiguated, so the union_int optimization won't be applied.

Later, it could be possible to make it possible to apply the union_int
optimization to these cases as well, but that would require more work,
possibly generating two variants of the function, one with and one
without decoding the first integer.

Add tests, splitting the UnionInt into two versions, one where the
optitmization is applied, and one where it's not.

Signed-off-by: Øyvind Rønningstad <oyvind.ronningstad@nordicsemi.no>