zcbor_encode.c: Fix error for big-endian architectures
Encoding small values with get_result_len() => 0, would read the wrong
memory byte to find the value to encode. This fix reads the byte that
corresponds to result_len == 1, which is required so the read isn't
out-of-bounds. result_len is 0 for these values, because they are encoded
in the header byte, so they don't use any additional bytes when encoded.

Add regression test.

Signed-off-by: Øyvind Rønningstad <oyvind.ronningstad@nordicsemi.no>