zcbor.py: Refactor the regex matching to fix label and remove all_types_regex
The label regex was broken, this fixes it. Also, adds ':' to the key matching.
This allows getting rid of the secondary key matching using all_types_regex.

Only the union matching ('/') needs all_type_regex, but it really only
needs the ones that match ranges, i.e. quotes and brackets. Refactor so it
only uses this.

Signed-off-by: Øyvind Rønningstad <oyvind.ronningstad@nordicsemi.no>