include: Rearrange and improve zcbor_decode.h and zcbor_encode.h
Attempt to make better groupings of functions, and add shared docs for
similar functions wherever possible. Also generally improve the docs.

Header file line count goes down significantly.

This is done in anticipation of adding a new class of functions, which
should be possible with minimal clutter in the files.

Signed-off-by: Øyvind Rønningstad <oyvind.ronningstad@nordicsemi.no>