zcbor.py: Fix a bug where some decoding function were not called
This is a bug in the union_int optimization. In maps, the optimization
refers to the key, but was propagated to the value as well, so that
in some cases, if the key was optimized away, the value was erroneously
optimized away as well.

Thanks to Pete Johanson for finding and debugging this issue!

Fixes #398

Signed-off-by: Øyvind Rønningstad <oyvind.ronningstad@nordicsemi.no>