zcbor.py: Fix a few bugs in code generation to do with dereferencing
- Two elements with different names referring to the same value were
  generated
- Sometimes a value was attempted dereferenced as result->somevalue,
  but it should have been just result.

This introduces the concept of delegated types, where the a type takes
on the C type of one of its children. This was handled via a shortcut
that failed for certain corner cases.

This also refactors the "skip" mechanism which was inconsistently
referring to itself or its parent, as part of the shortcut mentioned
above.

Add tests

Signed-off-by: Øyvind Rønningstad <oyvind.ronningstad@nordicsemi.no>