src: Use elem_state backup in relevant functions
I.e. in unions and repetitions. In both cases, it's ok to fail decoding
something because it's optional, and if the union element or repeated data
being decoded includes multiple map elements, backups are needed in case
flags have been set for some map elements, then fails later. In that case,
the previous map elements' flags must be undone, and a backup is the way
to keep the old state so it can be restored.

Signed-off-by: Øyvind Rønningstad <oyvind.ronningstad@nordicsemi.no>