zcbor_common.c: Address problem in zcbor_process_backup()
where a failure could leave the backup system in a bad state.
The elem_count check was placed later in the functions, so if it failed
some operations had already been performed, but not all.

As a fix, using the max_elem_count parameter to zcbor_process_backup*()
is deprecated. It should always be ZCBOR_MAX_ELEM_COUNT, and will be
removed in the future.
It was only used in one place, and this place now has its own check.

Signed-off-by: Øyvind Rønningstad <oyvind.ronningstad@nordicsemi.no>